requirements.txt
pyyaml>=6.0
jinja2>=3.1
weasyprint>=62.0
flask>=3.0
requests>=2.31
ruff>=0.15
pytest>=9.0
